Reorganize layout/: exactly two layouts (desktop + mobile)
The `layout/` directory had drifted into a junk drawer of
workspace components (TitleBar, TabBar, SidebarTree, the three
composables, ...). The user's mental model — and the right split
under style-guide.vue — is that `layout/` should hold only
*actual layouts*: the two ways the workspace can be assembled.

Layout splits like:

  layout/SidebarRight.vue
    Desktop: title bar across the top, editor + tabs as the main
    column, draggable Explorer sidebar pinned to the right edge.

  layout/Mobile.vue
    Smartphone-style responsive variant: title bar (taller),
    editor + tabs, bottom-nav bar pinned to the bottom, Explorer
    served as a bottom-sheet drawer when the user opens it.

`pages/layout.vue` picks one or the other via `matchMedia("(max-
width: 640px)")`; SSR defaults to desktop, the client switches on
mount and on subsequent resize.

Everything that used to clutter `layout/` moves into
`features/shell/`, colocated with the workspace state composables
that drive them (TitleBar, TabBar, EditorTab, BottomNav,
SidebarExplorer + SidebarTree + TreeNode + SidebarResizer, plus
useSidebar / useTabs / useDirs).

While moving things, also fix two coupled sidebar bugs:

* The drag-resize handle stopped tracking the cursor between
  SNAP_W (80) and MIN_W (200), because the move handler clamped
  the visible width to MIN_W. The splitter visibly froze on the
  200px line while the cursor kept moving — the "stuck white
  line" report. Collapse SNAP_W and MIN_W into one threshold
  (140) so the splitter follows the pointer 1:1 down to the snap
  point and the decoupled zone disappears.

* `is-active`/`resizing-sidebar` could stick after a pointer
  cancel because the cleanup path early-returned on lost capture.
  Move the cleanup into an `endDrag` handler that always runs on
  pointerup / pointercancel / lostpointercapture and tolerates
  capture-already-released errors.

* The `primitives/Sidebar.vue` mobile media query also stops
  honouring the `collapsed` class — that flag is a desktop "hide"
  toggle; on mobile the sheet's visibility is governed entirely
  by `open` (= `sidebar.menuOpen`).
