+++ apps/web/src/features/shell/TabBar.vue
